refactoring: context, reducer 관련 폴더 구조 개선
* context provider 와 useContext훅을 contexts 폴더로 이동시키킴
* reducer 는 reducers 폴더로 이동
