+++ frontend/guest-app/src/components/EmojiArea/EmojiArea.js
@@ -5,7 +5,7 @@ import EmojiPickerModal from "./EmojiPickerModal";
 import useCommonModal from "../CommonComponent/CommonModal/useCommonModal";
 import {socketClient} from "../../socket.io";
 import EmojiInsertButton from "./EmojiInsertButton.js";
-import useGlobalData from "../../models/GlobalData/useGlobalData.js";
+import useGlobalData from "../../contexts/GlobalData/useGlobalData.js";
 
 const RowWrapper = styled.div`
 	display: flex;
