+++ frontend/guest-app/src/components/QuestionContainer/EditQuestionInputDrawer.js
@@ -1,7 +1,7 @@
 import React from "react";
 import {socketClient} from "../../socket.io";
 import QuestionInputDrawer from "./QuestionInputDrawer.js";
-import useGlobalData from "../../models/GlobalData/useGlobalData.js";
+import useGlobalData from "../../contexts/GlobalData/useGlobalData.js";
 
 function EditQuestionInputDrawer({userNameRef, questionRef, toggleReducer}) {
 	const {event, guest} = useGlobalData();
