+++ frontend/guest-app/src/components/QuestionContainer/QuestionContainerTabBar.js
@@ -5,7 +5,7 @@ import Tab from "@material-ui/core/Tab";
 import Tabs from "@material-ui/core/Tabs";
 import gray from "@material-ui/core/colors/grey.js";
 import PropTypes from "prop-types";
-import useQuestions from "../../models/Questions/useQuestions.js";
+import useQuestions from "../../contexts/Questions/useQuestions.js";
 
 const RECENT_TAB_IDX = 1;
 const POPULAR_TAB_IDX = 2;
