+++ frontend/guest-app/src/components/SideMenu/MyQuestionButton.js
@@ -1,7 +1,7 @@
 import React from "react";
 import QuestionAnswerIcon from "@material-ui/icons/QuestionAnswer.js";
 import SideMenuItem from "./SideMenuItem.js";
-import useUIController from "../../models/UIController/useUIController.js";
+import useUIController from "../../contexts/UIController/useUIController.js";
 
 const MY_QUESTION_BUTTON_TEXT = "내 질문들";
