--- frontend/guest-app/src/models/UIController/useUIControllerReducers.js
@@ -1,17 +1,0 @@
-import useToggleReducer from "../Toggler/useToggleReducer.js";
-
-function useUIControllerReducers() {
-	const newQuestionInputDrawer = useToggleReducer();
-	const editQuestionInputDrawer = useToggleReducer();
-	const questionEditMenuReducer = useToggleReducer();
-	const myQuestionDrawerReducer = useToggleReducer();
-
-	return {
-		newQuestionInputDrawer,
-		editQuestionInputDrawer,
-		questionEditMenuReducer,
-		myQuestionDrawerReducer,
-	};
-}
-
-export default useUIControllerReducers;
