+++ frontend/guest-app/src/reducers/ToggleReducer.js
